functional_test/requirements.txt
python-decouple==3.4
google-auth==1.24.0
google-cloud-container==2.3.0
google-api-python-client==1.12.8
google-cloud-datastore==2.1.0
kubernetes==12.0.1
protobuf==3.14.0
requests==2.25.1
